// src/app/core/model/swagger-model/dayDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PhaseDTO } from './phaseDTO';

export interface DayDTO { 
    id?: number;
    name?: string;
    phases?: Array<PhaseDTO>;
}